Final improvements to the ORM refactor
